+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # CHANGELOG
 ### Version 10.3
+- Formula view table element for description and image re-written to div
+- Fix a font for the message returned when error occured during price get request from a supplier 
+- Further improvements in SDS generation
 
 ### Version 10.2
 - Minor improvements in finished product page