+++ releasenotes.md
@@ -1,3 +1,5 @@
 Whats New in v10.3
 --------------------------
+- Further improvements in SDS generation
+
 - For full details please refer to the CHANGELOG